move & split code
