+++ Balboa32U4/Arduino/FeedbackController/include/9axisIMU.h
@@ -1,0 +1,17 @@
+#include <LSM6.h>
+#include <LIS3MDL.h>
+
+LSM6 imu;
+LIS3MDL mag;
+
+float gx, gy, gz;
+float ax, ay, az;
+float mx, my, mz;
+
+void setupLIS3MDL(void);
+
+void setupLSM6(void);
+
+void setupIMU(void);
+
+void updateIMU(void);
