+++ Balboa32U4/Arduino/FeedbackController/include/conversion.h
@@ -1,0 +1,10 @@
+#define PI 3.14159
+
+//convert encoder val to wheel deg
+float en2deg(float encoder);
+
+//convert deg to rad
+float deg2rad(float deg);
+
+//convert rad to deg
+float rad2deg(float rad);
